requirements.txt
matplotlib==3.0.3
numpy==1.16.2
pyserial==3.4
